Update to 1.21.11

+++ Forge/src/main/java/net/potionstudios/biomeswevegone/forge/VanillaCompatForge.java
@@ -3,8 +3,8 @@ package net.potionstudios.biomeswevegone.forge;
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.InteractionResult;
-import net.minecraft.world.entity.npc.VillagerTrades;
-import net.minecraft.world.entity.npc.VillagerType;
+import net.minecraft.world.entity.npc.villager.VillagerTrades;
+import net.minecraft.world.entity.npc.villager.VillagerType;
 import net.minecraft.world.item.AxeItem;
 import net.minecraft.world.item.ShovelItem;
 import net.minecraft.world.level.block.*;
@@ -96,7 +96,7 @@ public class VanillaCompatForge {
             Int2ObjectMap<List<VillagerTrades.ItemListing>> trades = event.getTrades();
             BWGVillagerTrades.TRADES.get(event.getType())
                     .forEach((level, offers) -> {
-                        List<VillagerTrades.ItemListing> tradeList = trades.get(level.intValue());
+                        List<VillagerTrades.ItemListing> tradeList = trades.get(level);
                         tradeList.addAll(offers);
                     });
         }
